oci-umount: handle blank lines, wildcards (#756)
oci-umount in docker-1.12.6-55 introduces two wrenches: comment
and/or blank lines, and the use of '*' as a wildcard:

    # blah blah

    # blah
    /var/lib/docker/containers/*

Comments and empty lines are easy: ignore them.

Wildcards: will always be at the end, in the form 'dir/*',
meaning 'dir is OK but unmount anything underneath'. This
too is easy: strip off the asterisk but leave the trailing
slash. This will let us alert if we see dir/subdir mounted
but will not trigger on plain dir.

Signed-off-by: Ed Santiago <santiago@redhat.com>