dockerhelp: fix nonfunctional test
The dockerhelp subtest that was supposed to be checking
invalid commands ('docker --help --quiet', 'docker foobar')
was not actually doing anything useful. Remove nonuseful
code, replacing it with a simple check for output on
stderr: we don't check content, just check that something
was emitted on stderr.

Also, for proper help commands that should succeed,
make sure exit status is 0.

Signed-off-by: Ed Santiago <santiago@redhat.com>